+++ src/ekore/operator_matrix_elements/unpolarized/space_like/as3/aHg.py
